+++ src/modules/app/shells/CarbonShell.tsx
@@ -493,7 +493,6 @@ export default function CarbonShell({ children, kit, onKitChange }: { children: 
         show={showSelectionModal}
         onConfirm={handleSelectionConfirm}
         onCancel={handleSelectionCancel}
-        onExportToGithub={handleExportToGithub}
       />
       <GitHubExportModalWrapper
         show={showGitHubModal}